requirements.txt
fastapi[all]==0.104.0
langchain==0.0.324
lancedb==0.3.2
openai==0.28.1
python-multipart==0.0.6
uvicorn==0.23.2
pandas==2.1.2
tiktoken==0.5.1
passlib==1.7.4
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0